Improved some things with the subscription process.
Plans are now stored in the table 'subscription-type', which now has its
own service. Associated subscriptions with subscriptionTypes.

Added field customerId to subscriptions so we can save the Chargebee
customerId.

Removed free tier as a plan since that's just the default state of a
new user. If we really need to add it back in, it should be trivial.

Updated Chargebee redirect to include the subscription ID as a
parameter. This should be passed back to the server when Chargebee is
done so we can associate the call with which subscription has been
verified and paid for.

Calling subscription.create removes all existing unverified
subscriptions for that user.

Authenticating subscription calls, disallowing external update/patch/
removes for that service.

Added service to confirm subscription so Chargebee has something to
redirect users to. Only allowing get on that service.

User service now populates subscription and, if subscription exists,
populates subscriptionType as well in separate after() hook (this
latter part could definitely be done better).