Fixed issues with stopping and restarting video chat.
Tested audio and ensured it worked. Added volume controls for consumers and producers.

Found that new receive transport not transitioning to 'connected' connectionState after first call to .consume
didn't actually appear to indicate that it was not connected. Removed wait for that in subscribeToTrack and every
consumption seems to work.


Former-commit-id: 0430194f5744cc73f29d4dbc657ea41af6e6359f