Mediasoup on gameservers requires customization due to the need to open up UDP ports. The networking involved also requires that each gameserver only has a limited number of UDP ports, as each incoming UDP port can only point to a specific server.
Each gameserver will have to create and delete a service and ingress for itself, along with
patching the UDP configmap.

Started working on all the networking stuff. Ended up needing to pass app to the instantiation
of WebRTCGameserver so that the ServerTransport has it. Trying to import app from ServerTransport
was a circular dependency, so app wasn't fully exported and would show up early in initialization
as null.

Moved app/index to app.ts. Feathers-cli's generate command expects an 'app.ts' file, and I didn't
see a way to give it a different path.