Added some cleanup code to close leftover transports.
Sometimes, the client signals to the gameserver that it should close DTLS.
This was causing client disconnects to not trigger the normal transport
cleanup, which was leading to some weird behavior when that user
reconnected, namely video and audio streams that were muted.
There's now a call to closeTransport when dtlsState changes to 'closed'
and a call to clean up any transports that match peer, direction, and party
when a new transport is created.