Added subdomain information to instances in admin panel.
Made instance modal close if there are no more users on it.

Added some user-is-admin checks to backend services.
Updated handful of other calls to vanilla user.find()
to use 'action: layer-users' so as not to be blocked
by new admin check. All of these calls seemed to
be getting users on the current layer/instance anyway.

Fixed some styling issues.