Fixed issue with restarting audio not playing.
PositionalAudioSystem was not checking if there was a new consumer/track
for a given characterAudioStream, instead always assuming that the presence
of a track meant things were nominal and skipping adding a track to it.
Changed skip check to account for new consumers for the same character.

Added setting to enable/disable spatial audio. If on, spatial audio is used;
if off, PartyParticipantWindow audio is used. Volume controls need to be
re-enabled on PartyParticipantWindow if spatial audio is off.

User settings were never being created. They now are on identityProvider
creation, post user creation (which is currently never called, but just
in case), and on doLoginAuto if an existing user doesn't have settings
(mainly doing this to avoid having to reset the database and wipe
existing users).

Added some conditional checks for weird local circumstances where timing-out
connections were causing things that should be defined to be null/undefined.


Former-commit-id: 37b2fb6d628ab9d5aa3bcfbf4ad28b3978b49f6b