Updated instance-provision to ensure gameservers exist.
Encountered bugged situation when gameservers crashed. Instance
record is not removed if no new GS is spun up, but new connection
attempts to that location will try to go to the nonexistent server
specified in the DB.

Unhandled exceptions are only supposed to do synchronous calls, so
cleanup there is not a great option. Changed instance-provision find
to check if the existing GS that's been found actually exists, and
to clean it up if it doesn't and either find another one or allocate
a new GS if there aren't actually any in use for that location.