Fixed a bug with calls for WebRTCReceiveTrack().
On prod, the gameserver was sometimes crashing because handleWebRtcReceiveTrack()
in the gameserver transport was getting 'null' when finding a transport, and
transport.consume() was not conditional on transport being not-null.

Added a conditional check, and if it's null, just return an object with id: null,
so that the client will ignore that consume attempt.

Not sure why this is occurring, as that suggests the client's transport is crashing
or something, but at least this should prevent the whole GS from going down.