Fixed some build and deployment issues.
DRACOLoader.js and GLTFLoader.js were changed from 'export' to 'module.exports',
and for some reason this caused imports of them to not recognize what
was being imported. Reverted back to 'export'.

The check for whether a DB is running locally or as a Docker container should not
be run in production, as the DB is run externally. Added conditional statement
to skip that check in production.