Implemented working unreliable messaging.
Reconfigured transports and producers/consumers so that data can be sent
to/sourced from server's message queue as the single source of truth.

Only DirectTransports can be read from/written to on the server side.
WebRTCTransports, and other transport types, cannot.
Solution was to create a DirectTransport server side. Whenever a
DataProducer is created for a client, a DataConsumer is created on the
DirectTransport that consumes that DataProducer. This DataConsumer
is used to put the messages it's getting from the DataProducer onto
the server's incoming message queue.

The server sets up a single DataProducer on the DirectTransport when it starts.
This DataProducer is always sending out the messages popped off the outgoing
message queue. When a client signals it wants to create a DataConsumer, it's
consuming this outgoing DataProducer - all clients consume the same DataProducer.

Changed how data is sent/received by producers/consumers, as JSON.stringify()
does not properly transform ArrayBuffers. Had to do it via the follwing:

const stringified = JSON.stringify(Array.from(new Uint8Array(data)));
and
const decoded = Uint8Array.from(JSON.parse(message)).buffer;

Left data transfer as reliable messaging for now pending split in
which data should be sent reliably vs. unreliably.