Updated GS connection options to broaden A/V capabilities.
Previously, we were assuming all A/V was going to be on an instance,
and would either be instance-wide or party-based, so we just passed a
partyId around that would be 'instance' if it was for instance calls.

Added another parameter, relationshipType, that will be instance/group/friend/party.
partyId renamed to relationshipId, which is null for instance connections and
the id of the object being called otherwise.

Started updating instance fields; need to figure out how to have userRelationship
reference since just doing belongsTo(models.user_relationship) would create a
circular reference.


Former-commit-id: 9ec82b41691483d05cbc1e96d9e7b5be91df8121