Added login to Harmony and fixed OAuth bugs.
Decoupled ProfileMenu as much as possible from UserMenu so it
can be used elsewhere without requiring the full UserMenu.
Added ProfileMenu to Harmony.

Finally figured out OAuth bug, which was not actually anything to do
with OAuth, but rather a failure to update the JWT in localStorage
when doing loginUserByJWT().
Oauth strategies have been reverted to deleting guest identity-provider
and user if that OAuth user is already in the system, rather than the janky
patching of identity-providers that was a workaround.