Fixed a few issues relating to tokens expiring.
Server-side calls to authenticate tokens on GS (dis)connect throw errors if the token is expired.
On disconnect, would result in not updating tables for user disconnect, so made token expired error
instead use jsonwebtoken.decode to get the identity-provider ID and get the IDProvider manually.

Client was not properly handling invalid token on reAuthenticate. It now logs in as guest.

Removed throwing errors on GS (dis)connect, as that just causes the server to crash, which
probably doesn't help anything.