Split channels.ts and moved splits to gameserver/server.
channels.ts had logic for subscribing/unsubscribing users to Feathers channels on the API server
and handling socket connections/disconnects and associated instance updating on the gameserver.
Neither functionality was needed by the other server, so a channels.ts was created in each
server package that just had the things each needed. This removed engine as a dependency of
API server, which was causing tree shaking issues.

Also moved AvatarConstants from engine to common so that server-core doesn't depend on engine at all.
engine now depends on common for these constants.