Added modal that informs users of issues, also retries supplied action.
Modal is provided with title, body, countdown time, function to run, function parameters,
and optional function to run if the user closes the modal.
Title and body are displayed, modal counts down from time, then attempts to run the function
with the provided parameters and closes the modal. Closing modal kills the countdown timer.

instanceServerProvision and channelServerProvision updated to dispatch new events if their result
is null for either IP or port. Harmony and location listen for these events and open WarningRetryModal
on their reception. With Harmony, if the user actively closes the modal, it will end the current call.