Fixed issues with query param usage in login and other routing.
With removal of next, routing had been replaced with react-router-dom. This does not have
the query params on the 'match' prop that is passed to everything invoked with 'withRouter',
though, so all instances were replaced with alternate solution (calling new URLSearchParams
on useLocation().search, or window.location.search in one place that doesn't use components).