Switched some checks of NODE_ENV to check if Kubernetes is enabled.
There's a weird condition in our deployment where, inside the node process,
process.env.NODE_ENV is 'development' despite being set in multiple places
to 'production'. After a bug fix, this was causing SSL to be turned on for
the server and gameserver in production when it shouldn't have been.

Some of the checks for process.env.NODE_ENV === development, or !== production,
have been switched to check whether Kubernetes is enabled instead, which should
not have any oddities around it.