Updated kick handling to better notify user.
Being kicked from an instance due to a second connection wasn't being handled/displayed well.
Updated WarningRetryModal.tsx to allow for not retrying if noCountdown is passed in as true.
gameserver Kick emission now includes a message, which the client passes to the modal.

Updated SocketWebRTCClientFunctions.ts:leave to optionally be passed a 'kick' variable,
which if true will bypass trying to call LeaveWorld (it will not get a response, so no
reason to attempt it). Also updated the LeaveWorld request to use Promise.race() with a
10-second timeout so that it doesn't hang there.

Added a field 'left' to ClientTransport so that some disconnect handlers can not attempt
to re-connect; leave() is an intentional action that says the transport is done for good.