Fixed some issues with editor.
EditorContainer was trying to get the projectId from queryParams, but it's not in a query param, but
rather the end of the path. It also wasn't updating its own copy of state.pathParams when the projectId
changed after duplicating a scene (and possibly also updating the current scene).

There was also a bug where the metadata of a scene was being continually re-JSON.stringified, leading
to increasingly long stringified metadata. Made SceneNode.ts parse metadata down to an object and saved
in that format.