Fixed bug with starting consumers on initial click.
There was a gap in logic where, if a consumer existed before the initial click on a client's
screen, it was never getting resumed. The call to resume it was only coming when the AudioContext
was first created. An EngineEvent was added that AudioSystem now dispatches during startAudio(),
and PositionalAudioSystem now listens for that and tries to resume all suspended
queryResults.avatar_audio's.