Added analytics server, improved admin pages.
Added an analytics server. This server runs on its own and, at an
interval, collects the following metrics:
* Active Parties
* Users on world instances
* Users on A/V-only instances (channelInstances)
* Active Locations
* Active Scenes
* Active Instances

These are stored in a new 'analytics' table.

Instance records are no longer destroyed when the instance is shut
down. They are now marked as 'Ended'.

Added a new table 'instance_attendance'. It records the start and
end times of every occasion that a user was on an instance or channel.

Updated the root admin page to fetch these metrics and display them.
It also fetches the number of new users each day and the number of
active users per day. The latter is defined as a user that had at
least one instance_attendance record. (Note that it is possible for
more new users than active users, since the former is just how many
new records in the 'user' table there are for a given day, while the
latter is a user that went to an instance. You could log in and be
counted as a new user while not being counted as an active user).

Fixed pagination on admin pages. Previously, the table was just
paginating through the first page of results it got but couldn't
get any more pages.

projects.find now returns all scenes if the requesting user is an
admin, so /editor/projects will show all scenes for them.