Fixed some bugs in S3 project updating.
Made invalidation requests be for entirety of projects folder
(project-helper.ts:retriggerBuilderService) or a specific project folder
(project.class.ts:uploadLocalProjectToProvider). This saves a ton of invalidation
paths, which will cut down on costs.

Fixed a couple cases where params were null, which was causing params.sequelize to throw
errors due to reading a value from null.

Made createInvalidationCall not append a '/' to the start of each item if it already
has a leading slash.