Fixed bugs in chat and provisioning. (#4450)
Added association population to channel service. Updated channel.find
to just use the IDs of its findAll as part of super.find(params) so that
association population is always applied.

Removed search query param from friend finding. It wasn't used, and was causing
problems with new streamlined user.find.

Fixed a bug in ChatServive that was causing LOADED_CHANNEL channel patches to be
added as new, duplicate channels. Also made LOADED_CHANNEL remove existing instance
channel if one exists.

Fixed a bug in ChannelConnectionService that was matching any instance channel in
the channel list, rather than the one with the current instance's id. The above
ChatService fix should render this moot, but just in case, it should handle some
unforeseen circumstance.

Fixed issues with instance-provision being given IDs of invalid/ended/crashed
instances returning those bad instances' IP addresses as if everything was fine.
It now gets a free gameserver if the instance ended/is null, and if it appears to
be valid, performs gsCleanup() on it in case it had crashed. IP address of that instance
is only returned if, after all of this, the GS has been determined to be still running.